+++ lib/realtime_audio.dart
@@ -1,6 +1,8 @@
 export 'src/data/other.dart';
 export 'src/data/realtime_audio_echo_cancellation_state.dart';
+export 'src/data/realtime_audio_engine_health_event.dart';
 export 'src/data/realtime_audio_instance_response.dart';
 export 'src/data/realtime_audio_playback_clock.dart';
+export 'src/data/realtime_audio_playback_recovery.dart';
 export 'src/data/realtime_audio_state.dart';
 export 'src/realtime_audio.dart';